requirements.txt
# Core dependencies for voice cloning and translation
TTS>=0.17.0
torch>=1.13.0
torchaudio>=0.13.0
googletrans==4.0.0rc1
openai-whisper>=20231117

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-decouple>=3.8
numpy>=1.24.0
requests>=2.31.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
coqpit-config==0.2.1
coqui-tts==0.27.0
coqui-tts-trainer==0.3.1
cycler==0.12.1
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
grpcio==1.73.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
huggingface-hub==0.33.4
# idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.3
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.53.2
typeguard==4.4.4
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
